+++ test/dummy/app/controllers/user_tags_controller.rb
@@ -1,0 +1,10 @@
+class UserTagsController < ApplicationController
+  def index
+  end
+
+  private
+
+  def find_current_tenant_account
+    Account.find_by(domain: request.domain)
+  end
+end