+++ test/dummy/config/routes.rb
@@ -1,6 +1,7 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   resources :users
+  resources :user_tags
 
   namespace :super_admin do
     resource :dashboard, controller: :dashboard